+++ src/main/scala/io/frontroute/RouteLocation.scala
@@ -1,8 +1,11 @@
 package io.frontroute
 
+import io.frontroute.internal.HistoryState
+
 final case class RouteLocation(
   unmatchedPath: List[String],
-  params: Map[String, Seq[String]]
+  params: Map[String, Seq[String]],
+  state: Option[HistoryState]
 ) {
 
   def withUnmatchedPath(path: List[String]): RouteLocation = this.copy(unmatchedPath = path)