+++ src/main/scala/io/frontroute/ScrollPosition.scala
@@ -1,0 +1,6 @@
+package io.frontroute
+
+final case class ScrollPosition(
+  scrollX: Option[Int],
+  scrollY: Option[Int]
+)