+++ src/main/scala/io/frontroute/domext/HistoryWithTitle.scala
@@ -1,0 +1,10 @@
+package io.frontroute.domext
+
+import scala.scalajs.js
+
+@js.native
+private[frontroute] trait HistoryWithTitle extends js.Object {
+
+  val title: js.UndefOr[String] = js.native
+
+}